TODO Cleanup
Scrapped the idea of exporting settings.
If you want to share your settings, just screenshot app's UI haha

Scrapped the idea of convert for VV button, might implement someday but it'll be a secret setting, I have no idea where to put this button in the UI that'd make it relevant, it just seems like an out-of-place feature... or rather, bloat.

Deleted ideas to consider because I won't consider them anymore.
reasons were already written