Improve Previewer implementation
- Default values for resetting sliders are now read from a new defaults subclass
- Defaults subclass is used as a compass for Slider default values
- Fix Add Ambient Light toggle not persisting
- Updated UpdateUI method to account for Previewer, no more glitchy conflicts
This was done by hiding the vessels and setting an empty image every time the UpdateUI method is called, then making them visible again.

Previewer is no longer concerned with visibility, this makes visibility property a reliable way for handling this situation.

For better results, Previewers won't be set until after previous session's setting values are loaded and updated in the UI.

Previewr will be set only once in the UpdateUI method, the first time it is called.
Instead of making async methods and awaiting update UI, we run it at the end of UpdateUI, but only once using the RanOnceFlag class

Additionally, the reset button now resets the logs as well
This goes nicely with images suddenly getting wiped upon calling the updateclass method.