Update plans
reasons behind scrapped plans:

1.4 and 1.5 have been tested, everything's good.

Processing for other packs has been implemented, though still rough, needs work.

RTX Render distance override is not a good idea, the game doesn't like high render distances with RTX

All packs are indeed passed down to processors equally, each processor as "target files" to look for, if it finds them, it modifies them.
For Vanilla RTX it doesn't matter, but it allows tuner to process any pack with any combination of texture sets

as for having a worker that checks for Vanilla RTX Updates
well, It just doesn't make too much sense to trigger a remote update check besides one of the most used buttons in the app.
adding a dedicated button for it will make the ui too cluttered, more than it already is.

Besides, reinstall latest packs button always reinstalls the latest no matter what, it is robust, checks cache against remote.

and if Opus comes to be a thing in the future, I really don't have time to update all of the update checkers and reinstallers for it -- it remains a bit of an outlander