Update MainWindow.xaml.cs
The whole premise now is that the app opens QUICKLY and without any art at start up
these are ideas are just bad now.
why add an annoying artificial delay and bloat the code just so you can call update UI without worry in the background

It's very good as it is, the startup is clean. don't bloat it further.